[Bronze V] Title: 이상한 기호, Time: 172 ms, Memory: 17744 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -4,10 +4,10 @@ public class Main {
     public static void main(String[] args) {
         Scanner scan = new Scanner(System.in);
 
-        int N = scan.nextInt();
-        int M = scan.nextInt();
+        long N = scan.nextInt();
+        long M = scan.nextInt();
 
-        int result = (N + M) * (N - M);
+        long result = (N + M) * (N - M);
         System.out.println(result);
     }
 }
